+++ CHANGELOG.md
@@ -6,4 +6,5 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-- Package Updates
+- add process.env.NODE_PATH to resolve algorithm
+- package updates